Define model data classes and update api return type
